requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Werkzeug==2.3.7
PyMuPDF==1.23.7
fitz
pdf2image==1.16.3
Pillow==10.0.1
WTForms==3.0.1
python-dotenv==1.0.0